fixed WebAuthnE2ETest
